requirements.txt
flask
python-dotenv
librosa
numpy
joblib
tensorflow
edge-tts
datasets
soundfile
